Move templateName function to template maps model
