+++ lib/cartodb/controllers/template_maps.js
@@ -1,6 +1,7 @@
 var step = require('step');
 var assert = require('assert');
 var _ = require('underscore');
+var templateName = require('../template_maps').templateName;
 var CdbRequest = require('../models/cdb_request');
 
 function TemplateMapsController(app, serverOptions, templateMaps, metadataBackend, templateBaseUrl, surrogateKeysCache,
@@ -386,15 +387,4 @@ function ifInvalidContentType(req, description) {
     if ( ! req.headers['content-type'] || req.headers['content-type'].split(';')[0] != 'application/json' ) {
         throw new Error(description);
     }
-}
-
-function templateName(templateId) {
-    var templateIdTokens = templateId.split('@');
-    var name = templateIdTokens[0];
-
-    if (templateIdTokens.length > 1) {
-        name = templateIdTokens[1];
-    }
-
-    return name;
 }
