+++ lib/cartodb/template_maps.js
@@ -492,3 +492,14 @@ o.fingerPrint = function(template) {
     .digest('hex')
   ;
 };
+
+module.exports.templateName = function templateName(templateId) {
+    var templateIdTokens = templateId.split('@');
+    var name = templateIdTokens[0];
+
+    if (templateIdTokens.length > 1) {
+        name = templateIdTokens[1];
+    }
+
+    return name;
+};
